Refactor property module: Move EntityEditPagePresenter and EntityFormHelper to helpers
- Updated namespaces for EntityEditPagePresenter and EntityFormHelper to reflect their new location in the helpers directory.
- Created new EntityEditPagePresenter class to build XSL view-model payload for entity edit/view pages.
- Created new EntityFormHelper class to consolidate entity form workflows, including mapping, validating, persisting, and responding to entity form submissions.
- Adjusted references in EntityControllerTest to use the new namespace for EntityFormHelper.
